# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is an API server for asset management system.",
        "title": "Asset Management API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:6500",
    "paths": {
...
        "/asset/asset-list/create": {
            "post": {
                "description": "Create Asset Item record",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AssetList"
                ],
                "summary": "Create Asset Item record",
                "parameters": [
                    {
                        "description": "Asset List Request Body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.AssetLists"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.AssetLists"
                            }
                        }
                    }
                }
            }
        },
        "/asset/asset-list/list": {
            "post": {
                "description": "List Asset Items",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AssetList"
                ],
                "summary": "List Asset Items",
                "parameters": [
                    {
                        "description": "Asset List Request Body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.ListAssetReqDto"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.AssetListList"
                            }
                        }
                    }
                }
            }
        },
...
        "/base/asset-list/one/{id}": {
            "get": {
                "description": "Get one asset record by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AssetList"
                ],
                "summary": "Get one asset record by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "asset list ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.AssetLists"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.AssetListList": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "lists": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.AssetLists"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "dto.ListAssetReqDto": {
            "type": "object",
            "properties": {
                "assetCode": {
                    "type": "string"
                },
                "assetName": {
                    "type": "string"
                },
                "deptIds": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "placeIds": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "purchaseDates": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "typeIds": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "models.AssetListFiles": {
            "type": "object",
            "properties": {
                "assetId": {
                    "type": "string"
                },
                "base64": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "fileName": {
                    "type": "string"
                },
                "fileType": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.AssetLists": {
            "type": "object",
            "properties": {
                "accountCode": {
                    "type": "string"
                },
                "accountName": {
                    "type": "string"
                },
                "afterBeforeTax": {
                    "type": "number"
                },
                "assetCode": {
                    "type": "string"
                },
                "assetName": {
                    "type": "string"
                },
                "brandCode": {
                    "type": "string"
                },
                "brandName": {
                    "type": "string"
                },
                "chequeNo": {
                    "type": "string"
                },
                "cost": {
                    "type": "number"
                },
                "createdAt": {
                    "type": "string"
                },
                "deptId": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "includeTax": {
                    "type": "boolean"
                },
                "invoiceDate": {
                    "type": "string"
                },
                "invoiceNo": {
                    "type": "string"
                },
                "invoiceRemark": {
                    "description": "I",
                    "type": "string"
                },
                "maintenancePeriodEnd": {
                    "type": "string"
                },
                "maintenancePeriodStart": {
                    "type": "string"
                },
                "placeId": {
                    "type": "string"
                },
                "purchaseDate": {
                    "type": "string"
                },
                "remark": {
                    "type": "string"
                },
                "serialNo": {
                    "type": "string"
                },
                "sponsor": {
                    "type": "boolean"
                },
                "sponsorName": {
                    "type": "string"
                },
                "staffName": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "taxCode": {
                    "type": "string"
                },
                "taxCountryCode": {
                    "type": "string"
                },
                "taxInfoId": {
                    "type": "string"
                },
                "taxRate": {
                    "type": "number"
                },
                "typeId": {
                    "type": "string"
                },
                "unit": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "uploadAssetListFiles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.AssetListFiles"
                    }
                },
                "vendorId": {
                    "type": "string"
                },
                "voucherNo": {
                    "type": "string"
                },
                "voucherUsedDate": {
                    "type": "string"
                }
            }
        },
...
    }
}